ComputerVision/ImageSegmentation/BodySegmentation/requirements.txt
albumentations==1.0.3
catalyst==21.7
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.1
opencv-python-headless==4.5.2.54
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
tensorboardX==2.2
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
typing-extensions==3.10.0.0